# src/vmapi/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Virtual Machine Api",
        "title": "Virtual Machine Api",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/host/{host}": {
            "get": {
                "description": "Host Monitor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HOSTS"
                ],
                "summary": "Host Monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HOST",
                        "name": "host",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        },
        "/hosts/vm/{host}": {
            "get": {
                "description": "Host VM Detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HOSTS"
                ],
                "summary": "Host VM Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HOST",
                        "name": "host",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        },
        "/ips": {
            "get": {
                "description": "Get ip pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPS"
                ],
                "summary": "Get ip pool",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IPS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            },
            "post": {
                "description": "Create ip pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPS"
                ],
                "summary": "Create ip pool",
                "parameters": [
                    {
                        "description": "IP地址段",
                        "name": "net",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NET"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        },
        "/ips/{ip}": {
            "put": {
                "description": "Append a ip to ip pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPS"
                ],
                "summary": "Append a ip to ip pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a ip from ip pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPS"
                ],
                "summary": "Remove a ip from ip pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Err": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "message": {
                    "type": "object"
                }
            }
        },
        "model.IPS": {
            "type": "array",
            "items": {
                "type": "string"
            }
        },
        "model.NET": {
            "type": "object",
            "properties": {
                "network": {
                    "type": "string"
                }
            }
        },
...
        "model.Res": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "message": {
                    "type": "object"
                }
            }
        },
...
    }
}